Move `workspace` module into its own crate
