+++ crates/workspace/Cargo.toml
@@ -2,3 +2,31 @@
 name = "workspace"
 version = "0.1.0"
 edition = "2018"
+
+[features]
+test-support = [
+    "client/test-support",
+    "project/test-support",
+    "tree-sitter",
+    "tree-sitter-rust"
+]
+
+[dependencies]
+buffer = { path = "../buffer" }
+client = { path = "../client" }
+editor = { path = "../editor" }
+gpui = { path = "../gpui" }
+project = { path = "../project" }
+theme = { path = "../theme" }
+anyhow = "1.0.38"
+log = "0.4"
+postage = { version = "0.4.1", features = ["futures-traits"] }
+tree-sitter = { version = "0.19.5", optional = true }
+tree-sitter-rust = { version = "0.19.0", optional = true }
+
+[dev-dependencies]
+client = { path = "../client", features = ["test-support"] }
+project = { path = "../project", features = ["test-support"] }
+serde_json = { version = "1.0.64", features = ["preserve_order"] }
+tree-sitter = "0.19.5"
+tree-sitter-rust = "0.19.0"
