+++ crates/workspace/src/items.rs
@@ -1,11 +1,11 @@
 use super::{Item, ItemView};
-use crate::{project::ProjectPath, Settings};
+use crate::Settings;
 use anyhow::Result;
 use buffer::{Buffer, File as _};
 use editor::{Editor, EditorSettings, Event};
 use gpui::{fonts::TextStyle, AppContext, ModelHandle, Task, ViewContext};
 use postage::watch;
-use project::Worktree;
+use project::{ProjectPath, Worktree};
 use std::path::Path;
 
 impl Item for Buffer {
