+++ crates/workspace/src/sidebar.rs
