+++ crates/zed/Cargo.toml
@@ -36,6 +36,7 @@ rpc = { path = "../rpc" }
 sum_tree = { path = "../sum_tree" }
 theme = { path = "../theme" }
 util = { path = "../util" }
+workspace = { path = "../workspace" }
 anyhow = "1.0.38"
 async-recursion = "0.3"
 async-trait = "0.1"
@@ -83,6 +84,7 @@ project = { path = "../project", features = ["test-support"] }
 rpc = { path = "../rpc", features = ["test-support"] }
 client = { path = "../client", features = ["test-support"] }
 util = { path = "../util", features = ["test-support"] }
+workspace = { path = "../workspace", features = ["test-support"] }
 cargo-bundle = "0.5.0"
 env_logger = "0.8"
 serde_json = { version = "1.0.64", features = ["preserve_order"] }
