+++ crates/zed/src/menus.rs
@@ -1,9 +1,18 @@
-use crate::{workspace, AppState};
+use crate::{AppState, WorkspaceParams};
 use gpui::{Menu, MenuItem};
 use std::sync::Arc;
 
 #[cfg(target_os = "macos")]
 pub fn menus(state: &Arc<AppState>) -> Vec<Menu<'static>> {
+    let workspace_params = WorkspaceParams {
+        client: state.client.clone(),
+        fs: state.fs.clone(),
+        languages: state.languages.clone(),
+        settings: state.settings.clone(),
+        user_store: state.user_store.clone(),
+        channel_list: state.channel_list.clone(),
+    };
+
     vec![
         Menu {
             name: "Zed",
@@ -27,7 +36,7 @@ pub fn menus(state: &Arc<AppState>) -> Vec<Menu<'static>> {
                 MenuItem::Action {
                     name: "New",
                     keystroke: Some("cmd-n"),
-                    action: Box::new(workspace::OpenNew(state.clone())),
+                    action: Box::new(workspace::OpenNew(workspace_params)),
                 },
                 MenuItem::Separator,
                 MenuItem::Action {
